Move Checkout page containers
