+++ src/components/containers/Checkout/CartFooter/CartFooter.spec.js
