+++ src/components/containers/Checkout/CartItem/CartItem.css
