+++ src/components/containers/Checkout/CartItem/CartItem.spec.js
