+++ src/components/containers/Checkout/EmptyCart/EmptyCart.spec.js
